Extracts DirectionListAdapter and optimizes imports
