+++ app/src/main/kotlin/com/mapzen/erasermap/presenter/MainPresenterImpl.kt
@@ -86,7 +86,6 @@ public class MainPresenterImpl() : MainPresenter {
                 viewController?.hideRoutePreview()
                 destination = null
             }
-            viewController?.hideRoutingMode()
         } else {
             viewController?.shutDown()
         }
