+++ app/src/main/kotlin/com/mapzen/erasermap/view/ViewController.kt
@@ -17,7 +17,7 @@ public interface ViewController {
     public fun showActionViewAll()
     public fun hideActionViewAll()
     public fun collapseSearchView()
-    public fun showRoutePreview(feature: Feature?)
+    public fun showRoutePreview(feature: Feature)
     public fun hideRoutePreview()
     public fun hideRoutingMode()
     public fun showRoutingMode(feature: Feature)
